+++ mm/2s2h/Rando/ActorBehavior/EnGo.cpp
@@ -13,7 +13,7 @@ s32 func_80832558(PlayState* play, Player* player, PlayerFuncD58 arg2);
 void Rando::ActorBehavior::InitEnGoBehavior() {
     // Scripted Actors
     COND_VB_SHOULD(VB_EXEC_MSG_EVENT, IS_RANDO, {
-        u8 cmdId = va_arg(args, u8);
+        u32 cmdId = va_arg(args, u32);
         Actor* actor = va_arg(args, Actor*);
         MsgScript* script = va_arg(args, MsgScript*);
         Player* player = GET_PLAYER(gPlayState);
